requirements.txt

fastapi
uvicorn[standard]
torch
torchvision
pillow
